[ImageManager]  Solve BSoD when list is empty Hide "Delete" label when the list is empty.

+++ ViX/src/ImageManager.py
@@ -192,6 +192,10 @@ class VIXImageManager(Screen):
 				mtimes.append((fil, stat(self.BackupDirectory + fil).st_mtime))  # (filname, mtime)
 		for fil in [x[0] for x in sorted(mtimes, key=lambda x: x[1], reverse=True)]:  # sort by mtime
 			self.emlist.append(fil)
+		if len(self.emlist):
+			self["key_red"].show()
+		else:
+			self["key_red"].hide()
 		self["list"].setList(self.emlist)
 		self["list"].show()
 
@@ -287,12 +291,13 @@ class VIXImageManager(Screen):
 
 	def keyDelete(self):
 		self.sel = self["list"].getCurrent()
-		self["list"].instance.moveSelectionTo(0)
-		if self.sel.endswith(".zip"):
-			remove(self.BackupDirectory + self.sel)
-		else:
-			rmtree(self.BackupDirectory + self.sel)
-		self.refreshList()
+		if self.sel is not None:
+			self["list"].instance.moveSelectionTo(0)
+			if self.sel.endswith(".zip"):
+				remove(self.BackupDirectory + self.sel)
+			else:
+				rmtree(self.BackupDirectory + self.sel)
+			self.refreshList()
 
 	def GreenPressed(self):
 		backup = None
